+++ app/src/main/java/com/example/projectdva232v1/ui/ResultActivity/ResultFragment.kt
@@ -1,5 +1,8 @@
-package com.example.projectdva232v1.ui.example6
+package com.example.projectdva232v1.ui.ResultActivity
 
+import android.content.Intent
+import android.graphics.drawable.Drawable
+import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -7,15 +10,17 @@ import android.view.ViewGroup
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import com.example.projectdva232v1.MainActivity
 import com.example.projectdva232v1.R
 import kotlin.math.roundToInt
 
 // TODO: Needs renaming - alternatively copy class content and refactor to different type (AppCompatActivity?)
-class example6Fragment : Fragment() {
+class ResultFragment : Fragment() {
 
-  private lateinit var example6Model: Example6Model
+  private lateinit var resultModel: ResultModel
   lateinit var backgroundColor: View
   lateinit var activityLogo: ImageView
   lateinit var testScore: TextView
@@ -26,26 +31,25 @@ class example6Fragment : Fragment() {
   lateinit var activityName: String
 
 
+  @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
   override fun onCreateView(
     inflater: LayoutInflater,
     container: ViewGroup?,
     savedInstanceState: Bundle?
   ): View? {
-    example6Model = ViewModelProvider(this).get(Example6Model::class.java)
-    val root = inflater.inflate(R.layout.fragment_example6, container, false)
+    resultModel = ViewModelProvider(this).get(ResultModel::class.java)
+    val root = inflater.inflate(R.layout.result_layout, container, false)
 
-    // TODO: Use these values or somehow get values from previous test page in another way?
-    /*
     score = intent.getIntExtra(correctAnswers, Int)
     maxScore = intent.getIntExtra(totalAnswers, Int)
     activityName = intent.getStringExtra(activity)
-    */
 
     initView(root)
 
     return root
   }
 
+  @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
   private fun initView(view: View){
     backgroundColor = view.findViewById(R.id.resultBackground)
     activityLogo = view.findViewById(R.id.result_ActivityLogo)
@@ -53,20 +57,29 @@ class example6Fragment : Fragment() {
     testName = view.findViewById(R.id.result_TestName)
     closeButton = view.findViewById(R.id.result_CloseButton)
 
-    // TODO: Connect these setters to values from previous test page and relevant resources
-    /*
-      TODO: Create resources for category colors and icons (strings/drawables?)
-      backgroundColor.setBackgroundColor(/* Insert color resource here - depends on test-category */)
-      activityLogo.setImageResource(/* Insert image resource here - depends on test-category */)
+
+    var colorId: Int = resources.getIdentifier(activityName, "colors", context?.packageName)
+    backgroundColor.setBackgroundColor(colorId)
+
+    var iconId: Int = resources.getIdentifier(activityName, "drawable", context?.packageName)
+    var da: Drawable = resources.getDrawable(iconId, null)
+    activityLogo.setImageDrawable(da)
+
     testScore.text = calcPercentage().toString()+"%"
     testName.text = activityName
-    // Direct closeButton to TestSelector/other page?
-     */
+    closeButton.setOnClickListener {
+      sendToHome()
+    }
   }
 
   private fun calcPercentage(): Int {
     var result: Double = score.toDouble() / maxScore.toDouble()
     result *= 100
     return result.roundToInt()
+  }
+
+  private fun sendToHome() {
+    val intent = Intent(this.activity, MainActivity::class.java)
+    startActivity(intent)
   }
 }
