+++ app/src/main/java/com/example/projectdva232v1/ui/ResultActivity/ResultModel.kt
@@ -1,10 +1,10 @@
-package com.example.projectdva232v1.ui.example6
+package com.example.projectdva232v1.ui.ResultActivity
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class Example6Model : ViewModel() {
+class ResultModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is gallery Fragment"
