+++ app/src/main/res/layout-land/result_layout.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.example6.example6Fragment">
+    tools:context=".ui.ResultActivity.ResultFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
