+++ app/src/main/res/layout/result_layout.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     app:layout_insetEdge="none"
-    tools:context=".ui.example6.example6Fragment">
+    tools:context=".ui.ResultActivity.ResultFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
